requirements.txt
# Core dependencies (always required)
# None - uses Python stdlib only

# Optional dependencies (for email/calendar integration)
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Note: Install optional dependencies only if using OAuth:
# pip3 install -r requirements.txt
